chore: move conftest higher so that it can be reused by other tests, not just Playwright
