--- aiarena/core/tests/playwright/conftest.py
@@ -1,28 +1,0 @@
-import pytest
-
-from ...models import WebsiteUser
-from ..base import BrowserHelper
-
-
-@pytest.fixture
-def bh(live_server, transactional_db):
-    browser_helper = BrowserHelper(live_server=live_server)
-    return browser_helper
-
-
-@pytest.fixture
-def admin_user(db):
-    return WebsiteUser.objects.create_superuser(
-        username="george",
-        email="george@aiarena.net",
-        password="guest",
-    )
-
-
-@pytest.fixture
-def user(db):
-    return WebsiteUser.objects.create_user(
-        username="billy",
-        email="billy@example.com",
-        password="guest",
-    )
